chore: repo-shape fixes — split pr_status_cache tests, allowlist test-file growth
