+++ scripts/max_lines_allowlist.txt
@@ -9,8 +9,8 @@
 
 anyharness/crates/anyharness-contract/src/v1/events.rs 949 existing AnyHarness contract oversized file
 anyharness/crates/anyharness-contract/src/v1/sessions.rs 802 existing AnyHarness contract oversized file
-anyharness/crates/anyharness-lib/src/api/openapi.rs 853 existing AnyHarness oversized file
-anyharness/crates/anyharness-lib/src/api/router_tests.rs 1172 existing AnyHarness oversized test file
+anyharness/crates/anyharness-lib/src/api/openapi.rs 867 existing AnyHarness oversized file (+hosting schema registration)
+anyharness/crates/anyharness-lib/src/api/router_tests.rs 1302 existing AnyHarness oversized test file (+hosting PR-status route tests)
 anyharness/crates/anyharness-lib/src/domains/cowork/runtime.rs 1901 existing AnyHarness oversized file
 anyharness/crates/anyharness-lib/src/domains/mobility/service.rs 1522 existing AnyHarness oversized file
 anyharness/crates/anyharness-lib/src/domains/plans/service.rs 628 existing AnyHarness oversized file
